requirements.txt
python-docx
paddlehub
paddlepaddle
numpy
timeit